Re-engineer swizzle implementation to support iOS architecture...
Apparently, vararg IMP functions do not behave correctly on iOS, though
they work fine in the Simulator.  Arguments are lost unless the IMP
function explicitly specifies the arguments in its definition.

As such, I've had to re-engineer the solution to require the
PearlSwizzle invocation to explicitly enumerate the arguments and types
of the swizzled method in such a way that it can be used to create a
compile time function reference.  We use a macro around
imp_implementationWithBlock to achieve this.