Handle container growing specially for AutoresizingContainer because it is managed by Auto Layout.
AutoresizingContainer is not actually resized when it needs to grow, we
just pretend like it has been, because it will be resized later by AL,
at which point we will refit the content view to the AL sizing.

This is because the AL sizing is not necessarily going to be what the
container growing sets the size to be.