Different method of calling original swizzle implementation.
Temporarily restoring the original swizzle implemenation at the method
point for invocation has drawbacks; mainly that doing so causes the
swizzle implementation to be missing for all recursive calls while the
original is restored.

Now the original implementation is available through a C function call.
This also lightens the burden on the swizzle framework, speeding up the
process.